New edits in triDiag, Substitution classes
I added edits for functions in the TriDiagMatrixOps class, and the Substitution class. The file Vvalues has nothing in it, as I got stuck with the Dirichlet solver, and how to actually do that. That being said, there are a lot of notes marked by //TODO that didn't get completed before the due date.